Update SpringBootJoranConfigurator to use proper GraalVM format
Before this commit, the generated name for the inner class had the
wrong format <package>.<parent>.<child> (canonical name).
GraalVM expects $ to separate the parent from the inner class.

This commit updates SpringBootJoranConfigurator to generate
an appropriate format for a class name. Specifically, an inner class
should be separated by a dollar sign, not a dot.

See gh-44016

Signed-off-by: Dmytro Nosan <dimanosan@gmail.com>