Extract "second" qualifier constant in EntityManagerFactory config
Extracts the repeated string literal "second" used as a qualifier value
into a private static final constant. This change improves code
maintainability by centralizing the qualifier value and reduces the risk
of typos in multiple occurrences.

The constant is used in:
- @Qualifier annotations
- persistenceUnit parameter
- method names

No functional changes or behavior modifications are included in this commit.

Signed-off-by: JunggiKim <kimjg2477@gmail.com>
Signed-off-by: kimjg <k42ke@naver.com>