Improve handling of default values when printing the banner
Previously, default values for the following properties did not work:

- application.title
- application.formatted-version
- application.version
- spring-boot.formatted-version
- spring-boot.version

Instead of the default value, an empty string was used instead. For
example, ${application.title:Title} would be replaced with "" rather
than "Title" when the application title was unavailable.

This commit improves the ResourceBanner so that a placeholder's
default value is used. An empty string will still be used when no
default value is provided. For example, ${application.title} will
be replaced with "". As before, custom properties that are not
well-known will not be replaced at all. For example
${custom.property} will remain as-is in the printed banner when
the custom.property has not been set.

Fixes gh-44137