Support DataSource auto-configuration without spring-jdbc
Prior to these changes, auto-configured a DataSource required
spring-jdbc to be on the classpath even if the app made no use of any
of its features.

The changes largely remove the use of spring-jdbc during DataSoruce
auto-configure or disable some support (configuring an unpooled,
embedded database) in its absense. The extra unwrapping in
DataSourceBuilder has been removed as it appears to be redundant.
The existing test for deriving from an embedded database continues
to work without it.

Closes gh-43786