Align nativeTest to native to make it multi-modules friendly
This commit harmonizes the behavior of the native and nativeTest Maven
profiles. Previously, enabling the nativeTest profile would
automatically add an execution of the Spring Boot and Native Build Tools
plugins, even in a module that doesn't represent an application or a
Spring Boot-related module.

With this commit, the native testing feature is only enabled if the
plugins are defined in the project, either directly, or in a parent.

The documentation has been updated as the behavior of both profiles is
now much more consistent.

Closes gh-44696