Improve class conditions on auth server JWT auto-config
Prior to this change, introspection of the auto-configuration could
fail due to insufficient protection against missing classes.

This commit introduces an extra class-level check for Nimbus's
JWKSource which ensures that the auto-configuration backs off if
nimbus-jose-jwt has been excluded. It also introduces an inner-class
for the case where spring-security-oauth2-jose is not on the
classpath. This ensures that the method defining the jwtDecoder bean
does not cause an introspection failure when JwtDecoder is missing.

Closes gh-45177