Ensure consistent versions between dev and production classpaths
Prior to this change, versions in the dev and production classpaths
could differ. These differing versions could result in a transitive
dependency that should have been present in both development and
production only being present in the former. This would likely
result in failures at runtime.

This commit aligns the versions by adding constraints to the
production runtime classpath for each dependency in the runtime
classpath.

Closes gh-46043