Make Quartz DataSource init back off without spring-boot-jdbc
Previously, setting spring.quartz.job-store-type to jdbc assumed that
spring-boot-jdbc was on the classpath. If it was not, the context
would fail to refresh as DataSourceScriptDatabaseInitializer was not
available.

This commit updates the auto-configuration so that the DataSource
initialization will back off when spring-boot-jdbc is absent. This
allows an app to use Quartz without spring-boot-jdbc and
spring-boot-sql. This remains true even when using the JDBC job
store as long as something else has created the DataSource and
initialized its schema to contain the tables that Quartz requires.

Closes gh-46243