Guard ZipkinTracingAutoConfiguration against missing Encoding bean
This commit harmonizes the check for the presence of an Encoding bean.
Also, as this bean may be provided by another auto-configuration, this
commit orders them explicitly.

This means the dependency to spring-boot-zipkin has to elevate from a
test-only dependency to (at least) an optional dependency to validate
the links between the two auto-configurations.

Closes gh-46295