Exclude spring-boot-devtools from AOT processing with Maven
Previously, spring-boot-devtools was only excluded from native images
built with Gradle but not with Maven. This inconsistency meant that
Maven builds would include devtools in the AOT processing classpath
and in the native image, causing build failures.

This commit harmonizes the situation and excludes devtools in a similar
fashion with Maven.

See gh-46533

Signed-off-by: academey <academey@gmail.com>