Stop using custom name for elements resolution
Previously, when a property was customized with @Name, getter, setter,
field, and record component were looked up using the custom name, rather
than the property name. This lead to missing information as the custom
name is not meant to match the field name, for instance.

This commit keeps the custom name, but uses the property name as for
other cases.

See gh-46599

Signed-off-by: Bohdan Pryshedko <prishedko@gmail.com>