Using @Name annotation leads to a missed items in metadata.
When @Name annotation is used on a property, metadata like javadoc and deprecation are not extracted. The root cause is that @Name.value is used for locating corresponding code element, and it can find something only @Name.value is the same as the code element name. This commit fixes the root cause.

Signed-off-by: Bohdan Pryshedko <prishedko@gmail.com>