Fix race condition in OutputCapture
This race could occur because of the existing cache in OutputCapture: When data was written, while it was read, any following reading of data might miss the last output. The cache was not invalidated correctly in this case.

We are now versioning the cache entries to make sure this can't happen anymore.

Signed-off-by: Daniel Schmidt <daniel-github@ad-schmidt.de>