Polish ArchitectureCheckTests
Prior to this commit, it wasn't possible to write a test to verify that
a specific ArchRule applied only to a specific SourceSet.

This PR rewrites `ArchitectureCheckTests` to support different sources
and also adds functionality to configure the plugin.

With these changes, all tests now verify not only the
`checkArchitectureMain` task but also `checkArchitectureTest`.

See gh-46822

Signed-off-by: Dmytro Nosan <dimanosan@gmail.com>