Improve Hazelcast health indicator to check that Hazelcast is running
This commit improves the Hazelcast health indicator to fail if the
Lifecycle manager indicates that the instance is no longer running.

This is needed because when the default Out-Of-Memory handler in
Hazelcast detects a OOM error, and terminates Hazelcast, it does so on
the node but not via the Hazelcast lifecycle manager - and the results
is that it is still possible to successfully execute transactions in
Hazelcast after this termination. The health indicator should not report
UP after Hazelcast has detected an OOM and terminated itself.

See gh-46877

Signed-off-by: Tommy Karlsson <tommy.karlsson@leovegas.com>