maven-plugin aot test mojo: process aot when tests are built but not run
Currently, it is not possible to build the tests (with aot processing)
without running them. Maven provides two different configuration switches:

* skipTests builds, but does not run, tests
* maven.test.skip neither builds nor runs tests

Right now the spring boot aot processing will skip aot in both cases.
However, there's reasons to build but not run tests: in particular, this
breaks reproducible builds because the main run (with tests) has the aot
classes built, but then the reproducer run (skip running tests) does not,
leading to different jars.

In the case where tests are built but not run, we should still do AOT processing.

Signed-off-by: Steven Schlansker <stevenschlansker@gmail.com>