Move configuration property sections from buildSrc to spring-boot-docs
The configuration of the configuration property sections in the appendix
was hardcoded in the DocumentConfigurationProperties task in buildSrc.
When a change is made to the sections, this triggers an entire rebuild
due to how Gradle treats buildSrc.
This commit moves the configuration to the spring-boot-docs modules
build.gradle script where it can be declarative, avoiding unnecessary
full rebuilds when sections are modified.
Closes gh-23972

Signed-off-by: Kim Tae Eun <snowykte0426@naver.com>