Remove quoting for system properties
Before this commit, system properties with a value has been quoted all the time.
This is fine as long if the resulting string is passed into the RunArguments(String) constructor, which then parses it again into a String[].
This isn't fine if they are added to ProcessBuilder.command, because then quoting is unnecessary.

This commit also fixes the unnecessary Map -> String -> String[] parsing by adding the entries directly into the array.

Closes gh-46555