Limit Kotlin API docs to Kotlin code
Previously, the Kotlin API docs included Java APIs. This commit
corrects this by suppressing all Dokkatoo source sets other than
main (main in src/main/kotlin, javaMain is src/main/java).

Dokkatoo is sensitive to the order in which plugins are applied. This
commit adapts to the sensitivity by changing the order in which
the Dokatoo and Kotlin JVM plugins are applied. This prevents the
Dokatoo plugin for overwriting the configuration that's applied by
our Kotlin conventions.

Closes gh-47763